Separate JWS and JWE benchmarks
